+++ src/schematools/contrib/django/management/commands/import_schemas.py
@@ -5,7 +5,7 @@ from collections.abc import Iterable
 from pathlib import Path
 
 from django.conf import settings
-from django.core.management import BaseCommand
+from django.core.management import BaseCommand, call_command
 from django.db import transaction
 
 from schematools import validation
@@ -60,6 +60,16 @@ class Command(BaseCommand):
             schemas = self.get_schemas_from_url(options["schema_url"])
 
         with transaction.atomic():
+            if schemas and current_datasets:
+                # Check if there are datasets missing
+                missing_datasets = self.get_missing_datasets(current_datasets, schemas)
+
+                if missing_datasets:
+                    # If missing, delete these schemas from datasets_dataset table
+                    for dataset in missing_datasets:
+                        self.stdout.write(f"Missing dataset: {dataset.name}")
+                        call_command("remove_schemas", dataset.name)
+
             # Contains unsaved Dataset objects if dry_run.
             updated_datasets = self._run_import(schemas)
             if not updated_datasets:
@@ -76,6 +86,20 @@ class Command(BaseCommand):
 
             if options["create_views"]:
                 create_views(self, updated_datasets, dry_run=self.dry_run)
+
+    def get_missing_datasets(
+        self, current_datasets: dict[str, Dataset], schemas: dict[str, DatasetSchema]
+    ) -> list[Dataset]:
+        # Check if datasets_dataset table contains datasets not present in the schemas
+        current_names = set(current_datasets.keys())
+
+        # Get snake case names for schema objects
+        schema_names = {Dataset.name_from_schema(schema_obj) for schema_obj in schemas.values()}
+
+        # Datasets in current_datasets but missing in schemas
+        missing_datasets = current_names - schema_names
+
+        return [current_datasets[key] for key in list(missing_datasets) if key in current_datasets]
 
     def get_schemas_from_files(self, schema_files) -> dict[str, DatasetSchema]:
         """Import all schema definitions from the given files."""