+++ CHANGES.md
@@ -1,3 +1,8 @@
+# 2025-09-30 (8.4.0)
+
+* Delete datasets when dataset.json schema is deleted.
+* Added management commands to import scopes and publishers.
+
 # 2025-09-25 (8.3.6.post1)
 
 * Minor fix for error in scope filtering